Split Lens local store
