+++ service-worker.js
@@ -468,6 +468,7 @@ const APP_SHELL = [
   '/js/lens-url.js',
   '/js/lens.js',
   '/js/lens-local.js',
+  '/js/lens-local-store.js',
   '/js/lens-local-worker.js',
   '/js/lens-local-utils.js',
   '/js/lens-local-parsers.js',
