+++ tests/test-quality-guardrails.js
@@ -211,6 +211,7 @@ const uiWorkflowCheckJsModules = [
   'js/lens-actions.js',
   'js/lens-cache.js',
   'js/lens-library.js',
+  'js/lens-local-store.js',
   'js/lens-local-worker.js',
   'js/lens-local.js',
   'js/lens-page-shell.js',
