+++ tsconfig.json
@@ -125,6 +125,7 @@
     "js/lens.js",
     "js/lens-local.js",
     "js/lens-local-parsers.js",
+    "js/lens-local-store.js",
     "js/lens-local-utils.js",
     "js/lens-local-worker.js",
     "js/light-audit-ai-analysis.js",
